LaTeX Macros
These macros that can be used in the LaTeX code and refer to specified values in the Python code. Values are saved to a text file and called in the LaTeX document using a backslash (\) followed by its identifying four letter code (\ _ x _ _{}). Note that the Macro itself includes a space, so note needs to be left after calling the macro.